+++ src/utils/H2DE_vector2d.cpp
@@ -105,6 +105,23 @@ const bool H2DE_Vector2D<H2DE_Vector2D_T>::isNull() const {
 }
 
 template<typename H2DE_Vector2D_T>
+H2DE_Vector2D<H2DE_Vector2D_T> H2DE_Vector2D<H2DE_Vector2D_T>::rotate(const H2DE_Vector2D<H2DE_Vector2D_T>& center, float angle) {
+    constexpr float DEG_TO_RAD = M_PI / 180.0f;
+    float rad = angle * DEG_TO_RAD;
+
+    float cosA = std::cos(rad);
+    float sinA = std::sin(rad);
+
+    float dx = static_cast<float>(x - center.x);
+    float dy = static_cast<float>(y - center.y);
+
+    return {
+        static_cast<H2DE_Vector2D_T>(cosA * dx - sinA * dy + center.x),
+        static_cast<H2DE_Vector2D_T>(sinA * dx + cosA * dy + center.y)
+    };
+}
+
+template<typename H2DE_Vector2D_T>
 H2DE_Vector2D<H2DE_Vector2D_T> H2DE_Vector2D<H2DE_Vector2D_T>::getCenter() const {
     return *this / 2;
 }